+++ NEWS.md
@@ -2,10 +2,14 @@ The current version of the `MSEtool` package is available for download from [CRA
 
 # MSEtool 3.4.9999 - Development Version
 
+
+
+
 ## Fixes
 - fix to fleet-specific vulnerable biomass provided to `MMPs`
 
 ## New additions
+- historical MPAs now work in `multiMSE`
 - argument `parallel` in `runMSE` can takes a named list of MPs
 - two new functions `WHAM2OM` and `ASAP2OM` to import assessments into operating models
 - `Assess2MOM` to generate a multi-fleet operating model, either for a 2-sex population where recruitment is predicted from female spawning output, or a single-sex population.